backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
celery[redis]==5.3.6
redis==5.0.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
aiofiles==23.2.1
pikepdf==8.15.1
pdfminer.six==20221105
slowapi==0.1.9
pydantic-settings==2.2.1
pydantic[email]==2.7.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-cov==5.0.0
